+++ http/http_server_helidon/src/test/kotlin/com/hexagonkt/http/server/helidon/AdapterExamplesTest.kt
@@ -6,7 +6,6 @@ import com.hexagonkt.serialization.jackson.JacksonTextFormat
 import com.hexagonkt.serialization.jackson.json.Json
 import com.hexagonkt.serialization.jackson.yaml.Yaml
 import org.junit.jupiter.api.Disabled
-import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.condition.DisabledOnOs
 import org.junit.jupiter.api.condition.OS.WINDOWS
 
@@ -17,10 +16,11 @@ val formats: List<JacksonTextFormat> = listOf(Json, Yaml)
 internal class AdapterBooksTest : BooksTest(clientAdapter, serverAdapter)
 internal class AdapterErrorsTest : ErrorsTest(clientAdapter, serverAdapter)
 internal class AdapterFiltersTest : FiltersTest(clientAdapter, serverAdapter)
-internal class AdapterClientTest : ClientTest(clientAdapter, serverAdapter, formats) {
-    // TODO Fix this case
-    @Test @Disabled override fun `Form parameters are sent correctly`() {}
-}
+internal class AdapterClientTest : ClientTest(clientAdapter, serverAdapter, formats)
+internal class AdapterClientCookiesTest : ClientCookiesTest(clientAdapter, serverAdapter, formats)
+internal class AdapterClientHttp2Test : ClientHttp2Test(clientAdapter, serverAdapter, formats)
+internal class AdapterClientHttpsTest : ClientHttpsTest(clientAdapter, serverAdapter, formats)
+// TODO Add multipart test
 @DisabledOnOs(WINDOWS) // TODO Make this work on GitHub runners
 internal class AdapterHttpsTest : HttpsTest(clientAdapter, serverAdapter)
 internal class AdapterZipTest : ZipTest(clientAdapter, serverAdapter)
