+++ http/http_server_netty/src/test/kotlin/com/hexagonkt/http/server/netty/AdapterExamplesTest.kt
@@ -21,6 +21,10 @@ internal class AdapterBooksTest : BooksTest(clientAdapter, serverAdapter)
 internal class AdapterErrorsTest : ErrorsTest(clientAdapter, serverAdapter)
 internal class AdapterFiltersTest : FiltersTest(clientAdapter, serverAdapter)
 internal class AdapterClientTest : ClientTest(clientAdapter, serverAdapter, formats)
+internal class AdapterClientCookiesTest : ClientCookiesTest(clientAdapter, serverAdapter, formats)
+internal class AdapterClientHttp2Test : ClientHttp2Test(clientAdapter, serverAdapter, formats)
+internal class AdapterClientHttpsTest : ClientHttpsTest(clientAdapter, serverAdapter, formats)
+internal class AdapterClientMultipartTest : ClientMultipartTest(clientAdapter, serverAdapter, formats)
 @DisabledOnOs(WINDOWS, MAC) // TODO Make this work on GitHub runners
 internal class AdapterHttpsTest : HttpsTest(clientAdapter, serverAdapter)
 internal class AdapterZipTest : ZipTest(clientAdapter, serverAdapter)
