+++ http/http_server_netty_epoll/src/test/kotlin/com/hexagonkt/http/server/netty/epoll/AdapterExamplesTest.kt
@@ -24,6 +24,14 @@ internal class AdapterFiltersTest : FiltersTest(clientAdapter, serverAdapter)
 @EnabledOnOs(OS.LINUX)
 internal class AdapterClientTest : ClientTest(clientAdapter, serverAdapter, formats)
 @EnabledOnOs(OS.LINUX)
+internal class AdapterClientCookiesTest : ClientCookiesTest(clientAdapter, serverAdapter, formats)
+@EnabledOnOs(OS.LINUX)
+internal class AdapterClientHttp2Test : ClientHttp2Test(clientAdapter, serverAdapter, formats)
+@EnabledOnOs(OS.LINUX)
+internal class AdapterClientHttpsTest : ClientHttpsTest(clientAdapter, serverAdapter, formats)
+@EnabledOnOs(OS.LINUX)
+internal class AdapterClientMultipartTest : ClientMultipartTest(clientAdapter, serverAdapter, formats)
+@EnabledOnOs(OS.LINUX)
 internal class AdapterHttpsTest : HttpsTest(clientAdapter, serverAdapter)
 @EnabledOnOs(OS.LINUX)
 internal class AdapterZipTest : ZipTest(clientAdapter, serverAdapter)
